requirements.txt
pandas==2.2.1
tabula-py==2.9.0